Fix coverity issues

+++ samples/cpp/whisper_speech_recognition/audio_utils.cpp
@@ -87,7 +87,7 @@ ov::genai::RawSpeechInput read_wav(const std::string& filename) {
     }
 
     const uint64_t n =
-        wav_data.empty() ? wav.totalPCMFrameCount : wav_data.size() / (wav.channels * wav.bitsPerSample / 8);
+        wav_data.empty() ? wav.totalPCMFrameCount : wav_data.size() / (static_cast<uint64_t>(wav.channels) * wav.bitsPerSample / 8);
 
     std::vector<int16_t> pcm16;
     pcm16.resize(n * wav.channels);
